Splitted counters from sequentials tests. Updated travis.yml
