+++ BinPy/tests/sequential_tests.py
@@ -195,34 +195,3 @@ def test_TFlipFlop():
     assert test_TFF.state() == [0, 1]
 
     clock.kill()
-
-
-def test_BinaryCounter():
-
-    clock = Clock(1, 50000)
-    clock.start()
-    test_BinaryCounter = BinaryCounter(clock.A)
-    op = []
-    for i in range(5):
-        test_BinaryCounter.trigger()
-        op += test_BinaryCounter.state()
-
-    assert op == [0, 1, 1, 0, 1, 1, 0, 0, 0, 1]
-
-    clock.kill()
-
-
-def test_NBitRippleCounter():
-
-    clock = Clock(1, 50000)
-    clock.start()
-    test_NBitRippleCounter = NBitRippleCounter(3, clock.A)
-    op = []
-    for i in range(9):
-        test_NBitRippleCounter.trigger()
-        op += test_NBitRippleCounter.state()
-
-    assert op == [0, 0, 1, 0, 1, 0, 0, 1, 1, 1, 0,
-                  0, 1, 0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 0, 0, 1]
-
-    clock.kill()
